up #407  - fix option type for maxeval: it is an int

+++ feel/options.cpp
@@ -160,7 +160,7 @@ nlopt_options( std::string const& prefix )
         ( prefixvm( prefix,"nlopt.ftol_abs" ).c_str(), Feel::po::value<double>()->default_value( 1e-10 ), "NLopt objective function  absolute tolerance" )
         ( prefixvm( prefix,"nlopt.xtol_rel" ).c_str(), Feel::po::value<double>()->default_value( 1e-4 ), "NLopt variables  relative tolerance" )
         ( prefixvm( prefix,"nlopt.xtol_abs" ).c_str(), Feel::po::value<double>()->default_value( 1e-10 ), "NLopt variables  absolute tolerance" )
-        ( prefixvm( prefix,"nlopt.maxeval" ).c_str(), Feel::po::value<double>()->default_value( 30 ), "NLopt maximum number of evaluations" )
+        ( prefixvm( prefix,"nlopt.maxeval" ).c_str(), Feel::po::value<int>()->default_value( 30 ), "NLopt maximum number of evaluations" )
         ;
     return _options;
 }
